src/requirements.txt
dj-database-url==0.4.1
Django==1.9.6
djangorestframework==3.3.3
factory-boy==2.7.0
fake-factory==0.5.7
psycopg2
gunicorn==19.6.0
python-dateutil==2.5.3
six==1.10.0
